cosomis/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.1.1
cloudant==2.15.0
decorator==4.4.2
Django==4.1.1
django-bootstrap4==22.2
django-environ==0.9.0
django-unicorn==0.48.0
et-xmlfile==1.1.0
idna==3.4
mysqlclient==2.1.1
numpy==1.23.3
openpyxl==3.0.10
orjson==3.8.0
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.12
celery==5.2.7
django-celery-results==2.4.0
redis==4.3.4
boto3==1.24.61
botocore==1.27.61
django-storages==1.13.1
djangorestframework==3.13.1
drf-spectacular==0.23.1
django-extended-choices==1.3.3
